refactor: extract testing policy runtime state microcrate
