+++ Cargo.toml
@@ -72,6 +72,7 @@ members = [
   "crates/bitnet-testing-policy-interop",
   "crates/bitnet-testing-policy-runtime",
   "crates/bitnet-testing-policy-runtime-core",
+  "crates/bitnet-testing-policy-state-core",
   "crates/bitnet-startup-contract-core",
   "crates/bitnet-startup-contract-diagnostics",
   "crates/bitnet-startup-contract-guard",
